+++ src/main/java/de/maxhenkel/miningdimension/block/ModBlocks.java
@@ -1,7 +1,0 @@
-package de.maxhenkel.miningdimension.block;
-
-public class ModBlocks {
-
-    public static final BlockTeleporter TELEPORTER = new BlockTeleporter();
-
-}